rename: BoardDirection -> Direction 클래스 명 수정
