+++ src/main/java/chess/domain/piece/Pawn.java
@@ -1,7 +1,7 @@
 package chess.domain.piece;
 
 import chess.domain.board.ChessBoard;
-import chess.domain.position.BoardDirection;
+import chess.domain.position.Direction;
 import chess.domain.position.Position;
 
 public class Pawn extends Piece {
@@ -26,7 +26,7 @@ public class Pawn extends Piece {
     }
 
     boolean isNotReachable(Position start, Position destination, ChessBoard chessBoard) {
-        BoardDirection direction = BoardDirection.of(start, destination);
+        Direction direction = Direction.of(start, destination);
         //팀의 전진방향과 맞도록 위로 한칸 움직이는 것은 가능하다
         if (team.isTeamForwardDirectionsContains(direction) && start.squaredDistanceWith(destination) == 1
                 && start.isOrthogonalWith(destination)) {
